% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authall.R
\name{authall}
\alias{authall}
\title{authall
authenticate all for googleCloudStorageR and googleComputeEngineR}
\usage{
authall(project = "luxoydata",
  auth_json = "/home/neal/redryzen_authin/luxoydata.json",
  zone = "asia-east1-a")
}
\description{
authall
authenticate all for googleCloudStorageR and googleComputeEngineR
}
