% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depack.R
\name{depack}
\alias{depack}
\title{depack}
\usage{
depack(rpack = FALSE)
}
\arguments{
\item{rpack}{TRUE or FALSE only detach the rpack() packages (default FALSE to detach all non-base packages, TRUE to ONLY detach rpack())}
}
\description{
detach all non-base packages
}
