% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghl.R
\name{ghl}
\alias{ghl}
\title{ghl}
\usage{
ghl(opt = "outerHTML", ps, xpath = "//*")
}
\arguments{
\item{opt}{What result to return, one of outerHTML, innerHTML, href, text, tablelist, tabl, or other html attributes}

\item{ps}{page source of page to parse}

\item{xpath}{xpath of elements to search for}
}
\description{
Return specified data from a page source
}
