% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchRD.R
\name{launchRD}
\alias{launchRD}
\title{launchRD}
\usage{
launchRD(browser = "chrome", cver = "77.0.3865.40", adblock = FALSE,
  adblock_filepath = paste0(getwd(), "/adblock.crx"), notrack = TRUE,
  winsize = c(1280, 800), browserargs = c("nosand", "dispop",
  "disnot"), headless = F, port = 4444L)
}
\arguments{
\item{browser}{chrome or firefox}

\item{winsize}{set window size of bot}

\item{port}{port number}

\item{ext}{extension filepath to add}

\item{opt}{options to add (nosand=no sandbox; disnot=disable notifications; dispop=disable popups)}
}
\description{
Primary fb likes function
}
