% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit.R
\name{lit}
\alias{lit}
\title{lit(()}
\usage{
lit(browser = "chrome", port = 4444, headless = FALSE,
  foo_priority = c("standalone", "docker", "local"))
}
\arguments{
\item{browser}{chrome or firefox}

\item{port}{port number}

\item{headless}{logical (TRUE or FALSE, no quotes) browse headlessly?}

\item{foo_priority}{vector of "standalone","docker","local" (up to all three) in order of priority of use (i.e. if you want to use standalone, with docker as backup, use foo_priority=c("standalone","docker"))}
}
\description{
Primary selenium browser launcher
}
