% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_docker.R
\name{lit_docker}
\alias{lit_docker}
\title{lit_docker()}
\usage{
lit_docker(port = 4445, browser = "chrome", headless = TRUE)
}
\arguments{
\item{port}{port number}

\item{browser}{chrome or firefox}

\item{headless}{MUST BE =TRUE FOR USING DOCKER}
}
\description{
Primary selenium browser launcher
}
