% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcheck.R
\name{rcheck}
\alias{rcheck}
\title{rcheck}
\usage{
rcheck(package = "data.table", github_user = "NA")
}
\arguments{
\item{package}{name of package}

\item{github_user}{name of github user, for cran do "NA" or dont use argument}
}
\description{
Check for and install package if missing
}
