% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnord.R
\name{rnord}
\alias{rnord}
\title{rnord
reset nord connection}
\usage{
rnord(nordserver = "random", OS = "linux")
}
\arguments{
\item{nordserver}{default "random" which connects or reconnects to a random US server, optionally specify}
}
\description{
rnord
reset nord connection
}
