% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack.R
\name{rpack}
\alias{rpack}
\title{rpack}
\usage{
rpack(package_vector = "")
}
\arguments{
\item{package_vector}{String vector of packages to check for and install if missing}
}
\description{
Check for and install vector of packages if missing
}
