% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{rs}
\alias{rs}
\title{rs}
\usage{
rs(min = 1, max = 2)
}
\arguments{
\item{min}{min sleep time}

\item{max}{max sleep time}
}
\description{
System sleep for a random amount of time, to .001 seconds, between two limits min/max (seconds)
}
