% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchRD.R
\name{launchRD}
\alias{launchRD}
\title{launchRD}
\usage{
launchRD(timeout_pageload = "default", timeout_script = "default",
  timeout_implicit = "default", browser = "chrome", cver = "AUTO",
  adblock = FALSE, adblock_filepath = paste0(getwd(), "/adblock.crx"),
  notrack = TRUE, winsize = c(1280, 800), browserargs = c("nosand",
  "dispop", "disnot"), headless = F, port = 4567)
}
\arguments{
\item{timeout_pageload}{set a page load timeout in seconds, use a number without quotes as a number of seconds to set as timeout (default: 300 seconds)}

\item{timeout_script}{set a script load timeout in seconds, use a number without quotes as a number of seconds to set as timeout (default: 30 seconds)}

\item{timeout_implicit}{set a time to wait to find elements if not immediately available on page, use a number without quotes as number of seconds to set as timeout (default: 0, do not wait any time for elements not immediatly available)}

\item{browser}{chrome or firefox}

\item{winsize}{set window size of bot}

\item{port}{port number}

\item{ext}{extension filepath to add}

\item{opt}{options to add (nosand=no sandbox; disnot=disable notifications; dispop=disable popups)}
}
\description{
Primary selenium browser launcher
}
