% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_local.R
\name{lit_local}
\alias{lit_local}
\title{lit_local()}
\usage{
lit_local(port = 4444, browser = "chrome", headless = FALSE,
  retry_max = 2)
}
\arguments{
\item{port}{port number}

\item{browser}{chrome or firefox}

\item{headless}{logical (TRUE or FALSE, no quotes) browse headlessly?}
}
\description{
Primary selenium browser launcher
}
