% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_standalone.R
\name{lit_standalone}
\alias{lit_standalone}
\title{lit_standalone()}
\usage{
lit_standalone(browser = "chrome", port = 4444, headless = FALSE,
  retry_max = 2)
}
\arguments{
\item{browser}{chrome or firefox}

\item{port}{port number}

\item{headless}{logical (TRUE or FALSE, no quotes) browse headlessly?}
}
\description{
Primary selenium browser launcher
}
